Covid programs splitted between fed and qc working
