--- srd/covid/__init__.py
@@ -1,1 +1,0 @@
-from .programs import programs, policy
