--- srd/delete.txt
@@ -1,1 +1,0 @@
-Hello World!
