+++ srd/quebec/__init__.py
@@ -1,2 +1,3 @@
 from .template import template, create_return
+from .covidqc import programs, policyqc
 from .forms import form, form_2016, form_2017, form_2018, form_2019, form_2020, form_2021
