+++ srd/quebec/params/covid.csv
@@ -1,0 +1,7 @@
+variable;value;type;definition;source;comments
+begin_april;3;int;;;programs begin in april 2020 (march 15th in reality)
+iprew_min_inc_work;5000;float;;https://www.revenuquebec.ca/fr/salle-de-presse/actualites/details/167331/2020-04-03/;
+iprew_max_inc_tot;28600;float;;;
+iprew_monthly_amount;400;float;;;
+iprew_max_earn;2500;float;;;max monthly earnings allowed
+iprew_max_months;4;int;;;iprew lasted 4 months
